Rename gcp-flex-start CLI subcommand to gcp
The name was misleadingly specific — the GCP provider already supports
multiple provisioning models (FLEX_START, SPOT, STANDARD) via the
--provisioning-model flag. Also pin venv to python3.12 in Makefile.
